Don't try to mock UserPreferences.
The test creates a mock version of an instance of UserPreferences.  But, it doesn't do anything with it.  The FileEpisode class acquires the singleton instance of UserPreferences, within its own source code.  Creating a mock instance does nothing to affect this.

If the static method, UserPreferences.getInstance(), could be mocked, that could work.  But, Mockito explicitly does not mock static methods.

The best solution, then, is to use a genuine instance of UserPreferences.  And instead of providing a mock value to be returned upon "get", we actually set that value, and then the get returns it.

One concern might be that the "set" would change the user's actual stored preferences, but "store" is only called when the UI dialog is dismissed.  Setting the value within the test does not affect the saved value.

Also, since our expected values of moved files does not include a directory, be sure to set "moveEnabled" to false.