Test for resolution.  Strip only junk.
The way removeLast was written, it removed the offending string, plus everything after it.  I don't think that was the intent.  It never broke anything in the test suite until now, though.  I'm adding a test to see that the screen resolution is parsed properly from the filenames, and we have one example where the resolution occurs after the "junk".  This fixes it.