Create FileEpisodes with just a file or filename.  Parse filename afterwards.
We should be able to have an object representing the episode immediately, as
soon as we know which file it refers to.  Parsing the filename can come later
(even if it should be immediately afterwards).

There are a few ramifications to this.  One is that the fields in FileEpisode
are no longer final.  But that actually fits the model better.  This program
is a renamer.  The information in the filename may change.  Currently, we're
still not updating the information after a change, but we should, and this
facilitates it.

Also, we used to only create a FileEpisode if we could parse the filename.
That's no longer the case.  Again, this change doesn't do much on its own,
because we still "reject" FileEpisodes with unparsable filenames.  But
in the future, we'll have more opportunity to work with them.  (For example,
we may want to present the user with the opportunity to help us figure out
what the filename means.)

I'm also changing the name of some of the member variables and their associated
getters and setters, to reflect that the season, episode, and resolution we have
here is based purely on the filename, and is not necessarily verified.