Parse dates inside Episode class, using static final DateTimeFormatter.
I had been doing the parsing inside TheTVDBProvider, because the old date formatter was not threadsafe, so we created a new instance in each call to getListings.  Now we create one single instance in Episode, and now we simply pass the text value from the XML to the Episode Builder for all fields.