Recognize underscore as a separator character.
The \W character class matches anything besides a letter, a number, or an underscore, and was used to recognize separators in filenames.  So, fields could be separated by dots, spaces, hyphens, etc.  But it didn't work to have an underscore separator, because the underscore is part of \w, meaning not part of \W.  So, we can't really use a built-in class.  Use a more explicit character class: [^a-zA-Z0-9]