Add Apply button to the Preferences dialog.
This allows the user to see the effects of their changes (at least, those that have visible effects), without dismissing the dialog.

Also, "Apply" is now the default (i.e., what happens when the user hits enter).

This change also repositions where the dialog box opens.

Previously, the dialog box had "Save" and "Cancel" buttons.  If you saved, the dialog box closes immediately after saving.  And so if you hit the other button, it implies that you had not yet saved (because if you had, the dialog would have closed).  So, it allowed you to throw away any edits you may have made.

With "apply", that would be a lot more difficult to implement, and it would be confusing, to boot.  So, we're just not doing it.  There's no more "Cancel".  There's just "Apply" (update all the attributes in the running application, and keep the dialog box open), "Save" (update any changes that hadn't yet been "applied", save the changes to the XML, and close the dialog box), and "Close" (close the dialog box, discarding any un-applied changes, but keeping any that have been applied).

The buttons could still perhaps be improved.

Also try to keep lines and methods to a reasonable length, here.