Add caching of files downloaded from tvdb v1
This is a development feature which never was made production-ready.  Since the v1 API is being phased out, it will not be.  Preserve the code as an intermediate commit.

When we want to look up either options for a given show name, or listings for a given show, now we first check the tvrenamer directory on the local disk for an already-downloaded result.  If it's not found there, then we go out to the provider and download the results, and store them in the tvrenamer directory.

Then, of course, next time, the results are sitting there waiting for us.  So, this implements a basic cache.

I think this could be useful to "ordinary" users, as it could speed things up in some cases.

But, really, that's not the reason for this commit.  It's much more useful for developers.

For one thing, it allows us to be good citizens of thetvdb.com.  I don't hear them complaining, and I assume they're set up to handle well more traffic than a few developers can possibly send, but still, we shouldn't be hitting their server when it's really unnecessary.  Obviously part of fully testing the application is that the download is working, but most of the other testing does not need to re-download data every time it's run.

Though, lately, their site has been down more frequently, and having the cache allows us to continue development even if they're offline.

It's also useful to have the XML around to be able to examine it manually.  The XML we get from the TVDB has no newlines.  I added the "INDENT" property to the output transformer, which makes it a lot better.  But it doesn't exactly indent the whole file as you might think.  Basically it gives each episode a single, often very-long line.  Still better than one line for the entire listings document.

And, skipping the download can make the edit-build-test cycle faster.

This is not a proper cache, as of now.  The files never expire.  That's not really appropriate for end users, but in the future, I could adapt it, or just make it optional (and off by default, but on by default for certain tests), or I could look into integrating an existing package that provides a proper cache, or...

It definitely makes sense to reload the episode list for current shows, and so, this commit temporarily breaks that.

For episode listings for shows that are off the air, it could be better to avoid downloading every time.  (Other details do change, even if there are no new episodes, but we don't currently use any of that data.)

For the show options, I actually think keeping things local will be the way to go, going forward.  But more details on that in future commits.

Again, this commit makes the branch unsuitable for general release.  The entirety of the v1 functionality will be removed for the next release.