Reinstate unused code.
For the 1.0b1 release, we removed any code that wasn't used, including a lot of code that was truly "dead".  But there was other code that was in place with the intent to use later.  I removed it for the release, but now I'm reinstating it.  Obviously, if it's still not used at the time of the next release, we'll remove it again.  :-)