Look up icon in UIUtils
Since we had essentially the same code in UIStarter and AboutDialog, implement it in UIUtils and kind of "cache" the result.

As a nice side effect, now neither UIStarter nor AboutDialog depend on java.io at all.

This also uses try-with-resources and some slightly modified logic to make sure the input stream is closed even in the case of a checked exception.  Findbugs had warned that we previously weren't doing that properly.