Repurpose GlobalOverrides
It was used at the renaming step.  I've deleted that usage.  Now I'm using it to override queries.  It's kind of the opposite usage, and the values get reversed.  It had been used to avoid writing out a superfluous year to the filenames; now use it to make sure the year is included for the query.