Save and load show and episode information to cache.
Add a second, better cache.

We already were caching the results we get from TheTVDB.  But these are large and very slow to parse, and contain more information than we need.  On a side note, they also are hard for a human to inspect because they lack indentation.  The files we generate here are much faster, because they map precisely to an instance of a class.  And they're easier to look at, too.

Just like with the TVDB cache, this is an incomplete implementation.  Nothing ever expires or is evicted.  But that's ok, because it's a development-only feature for now.  This would need to be improved before being merged to master.