Use TheTVDB Swagger API client.
The previous commits added all the functionality, but didn't use it.  This switches the application to use the Swagger class.

In the v1 file, make class no longer public.  In the v1 Test file, comment out the @Test annotations on most tests.  We are making the v1 code obsolete with this commit, but keeping it around just in case.  We'll do one quick test as a sanity check, but we don't need to run the full test suite any more.  The old functionality could easily be switched back on, if there's ever a reason to.