Add TheTVDB Swagger API test.
The Swagger test is pretty much an exact copy of the v1 test file.  Actually we had a few episodes commented out in v1, and I uncommented them in the Swagger file.

Comment out most of the @Tests in the Swagger file, because it's not being used yet.  This commit just adds it.