Change master branch to 0.8 sources
This is kind of a clunky and unfortunate thing to do.  We had been using the master branch for the 1.0 beta release.  Change it to be in synch with the 0.8b2 sources, by just committing those sources on top of master.

No commits have been reverted or removed.  All the history is still there, and will continue to be there, and those commits can be re-added in the future.