Files with different extensions are not conflicts
Address issue #259.

This bug came about due to the fix for #181 as well as #148.

The user may have multiple files that refer to the same episode.  We definitely do not want to cause them to overwrite each other.  Moreover, I think we really do not want to have duplicate episodes.  Most software will play files in order, so you can start with episode 1, and it will just play all the episodes you have, in order.  You presumably do not want to see the same episode twice.

For that reason, I specifically did not consider the file extension.  Whether it was .avi, .mp4, .mpg, whatever, you really only want each episode one time.

However, this neglected to consider auxiliary files, like subtitle files.  These should not be considered a conflict; they are complimentary to the episode file.

Moreover, I did not give users any option to control this behavior, although I intended to.

This commit changes it to consider file extension, and therefore, only consider a file a conflict if the two files truly want to have the exact same pathname.  When that's the case, the "versions" behavior is still in effect.

But when they have different file extensions, they will both be placed in the same spot.  This fixes the bug where the subtitle files were moved to the wrong place, and also gives what is likely the expected behavior in the case of identical episodes in different formats.

I still think this is not the optimal behavior, so I will re-introduce the ability to have only one episode file per episode, in a future commit, in a configurable way.