Fix typo and move constant strings
Noticed a typo in the dialog box: it said "There is a no new version available"; that " a " was ungrammatical.

I thought I'd find the text in Constants.java, but it wasn't there.  There were still a lot of constant strings in AboutDialog.java.  So, move them all into Constants.java, while fixing the one typo.