Allow sorting by non-text columns
In sortTable, to compare the values of the cells, we were using getText(column).  That worked fine for columns that have text in them, but didn't work with columns that had a check box or an icon in them.  Make a generic method that can get a String value for any column, and use that instead.