Don't rely on Show subclasses
We have a few different flavors of "Show".  It could be that we found the show in the provider's db, or that we have created a stand-in show, or that we failed to find the show.  These are implemented using subclasses, and then users check "instanceof" to see what it is.  For now, we're keeping the subclasses, but relegating it to an implementation detail by not relying on it directly, outside of the Show class.

In TheTVDBSwaggerProviderTest, add a test that the Show should not be a "failed show".  Previously we didn't bother with this, because we knew that a failed show was a subclass of a "local show", and we already verified that it wasn't a local show.  Since we're trying to be less reliant on clients knowing implementation details, add the extra (even though currently redundant) check.

Now can make LocalShow package-private.