Don't set FileEpisode path if value is null
In the code where we have tried to move a file, we have an intended destination, and then an "actual" destination.  Obviously if the file is moved, we expect it to be to the destination we specified, but there can be different ways to express the same path, so we use the "actual" destination to tell the FileEpisode where it lives now.

But if there's a problem moving the file, actualDest is null.  We should never set the Path of a FileEpisode to null, and anyway, we shouldn't be resetting the Path if we didn't move the file.  So simply do a null check.