Make code that doesn't rely on instance data, static
Make methods classForName, convertToLong, and registerName be static, as they doesn't rely on any instance data.

Make new static method initializeClass, which does what the beginning of initialize() used to do.  Call initializeClass() from initialize().

Not sure how good an idea this really is.  It gets findbugs to stop complaining.  Not sure why so much of this was non-static in the first place.

May not actually merge this into master.