Appease IntelliJ spell checker
Want to run IntelliJ spell checker on the code base.  Previous commit corrected some actual typos.  This commit is just a few things where the spell checker doesn't pick out the words correctly.  There was really nothing wrong with the way things were, in these cases, but it's nice to get the tool to find the minimal amount to complain about.