Small improvements to ShowName
We have procedures in place to try to make sure that, during a single run of TVRenamer, we don't look up options for a given show name more than once.  But it depends on which methods you call.  With TheTVDBProviderTest, we go directly to TheTVDBProvider, and so it's possible to call getShowOptions on the same ShowName more than once.  In that case, we get a bunch of "duplicate" options, because we're adding each option to a List that already contains each option.

Add a new method to clear the show options, and always call that method from TheTVDBProvider.getShowOptions().

Add a toString method to the ShowOption inner class.

In QueryString inner class, when we call setShow and decide that the best match is the one we had already matched, don't give an "info" message, just a "fine"-grained one, and change the wording slightly.