Clear episodes before adding episode infos
We have procedures in place to try to make sure that, during a single run of TVRenamer, we don't look up listings for a given show more than once.  But it depends on which methods you call.  With TheTVDBProviderTest, we go directly to TheTVDBProvider, and so it's possible to call getShowListing on the same show more than once.  In that case, we get a bunch of "duplicate" episodes, because we're adding each episode to a List that already contains each episode.

Add a new method to clear the episode infos, and always call that method from TheTVDBProvider.getShowListing().