Clear show options from ShowName before adding them
We actually go out of our way to re-use ShowNames and not have to look them up multiple times.  But for testing, we may look them up more than once.  In any case, if we do collectShowOptions, the options in the NodeList are the only options that should be allowed, so it makes sense to make sure that we start clean before adding the options.

As part of clearling away the ShowName, clear away any QueryStrings and any Shows that might have been associated with it.