Improve exception in getShowInstance
Previously, we just logged a message, saying that the ID could not be parsed as an integer.  That's not the only type of error that could occur.  Moreover, we should now consider it a little bit more severe.  With the TVDB v1 API, the ID was a String that had to be parsed, and who knows what they might give us?  With the new API, it is an integer to begin with, so if it can't be parsed, something really went wrong.