doesn't make sense to have move and rename both disabled
Make sure that if move is disabled, that rename is enabled.

In the preferences dialog, if move is disabled, do not allow the user to disable rename.  If rename is disabled, do not allow the user to disable move.

With the change to the preferences dialog in place, the user would not be able to turn both options off.  But, it is theoretically possible that they turned them both off in a previous release of TVRenamer, and saved them out that way, and then we would load them back in that way.  So, do a check when we load UserPreferences, and if they're both disabled, enable rename, and store the prefs.

By the same token, add code to setMoveEnabled and setRenameEnabled to prevent them from both being disabled, even though there should be no way to reach that code in the application.  Just in case.