Make all parameters final in ResultsTable.java
Replace ProgressObserver.cleanUp method with finishProgress

New method takes a boolean to say if it succeeded or not.  It performs an extra null check on the label.

In FileCopyMonitor, always set the display variable (don't wait until initialize()).

With these two changes, it's safe to call finishProgress even without calling initialize.

Add user preference deleteRowAfterMove

The intent is that when deleteRowAfterMove is set, the UI will automatically delete rows after the files have been successfully moved/renamed.

Add a checkbox to the preferences dialog for this.

Add finishMove method

Now from FileMover, we call finishProgress() in the FileCopyMonitor even if we weren't using a progress label.

From finishProgress(), we now call finishMove() in the ResultsTable.

finishMove() will, if deleteRowAfterMove is set and the move succeeded, delete the table item.

Change listingsComplete to return the number of options