Move column header texts into Constants
Don't say "selected" when we mean "checked"

The table actually allows you to select rows.  And yet, that selection has nothing to do with what happens when you press the button to move/rename files.  Change labels to be a little clearer (even if a bit more awkward).

Change constant name from SELECTED_COLUMN to CHECKBOX_COLUMN

Change local variable from selectedColumn to checkboxColumn

Update comment and log message to not refer to "selected" to mean "checked"

Change variable name from renameSelectedButton to actionButton

It doesn't operate on "selected" items, but "checked" items.  And it doesn't necessarily change the file name, it might just move the file.

Inline tableItemDownloaded

tableItemDownloaded was a private method that was only called from one place, so move its code into its caller.

Set focus after pressing button.

By clicking the "rename/move" button, focus is transferred to the button.  To allow the user to keep using keystrokes on the table, give the focus back to the table after all the moves are enqueued.

Improvements around sortTable

Change sortTable to not take columnNum.  Previously calling sortTable required passing in columnNum.  Our idea of columnNum was, frankly, an assumption.  As far as I know, it was always right, but it still wasn't being done right.  Instead of passing it in, we will look it up properly inside sortTable.

Add Javadoc comment for sortTable (even if it's private).

Pull setColumnDestText out of setupColumns.  After the columns are set up, set a default sort column and direction for the table.