Make new column sort always sort up
Change the interpretation of "up", to match what the OS X Finder does.  It uses the up arrow for A-Z and the down arrow for Z-A.  We had been doing the opposite.  Now we do it the same way as the Finder.

And, any time sortTable is run, we start with the assumption that we are sorting "up".  But if the previous sort column is the same as the current sort column, we sort in the reverse direction as was previously sorted.