Add originalBasename variable
The originalBasename is what you get when you take original file path, remove the directory, and remove the file suffix.

So, for "Movies/TV/Lost-S01E04.avi", the originalBasename is "Lost-S01E04".

Contrast this with baseForRename; that is the basename of the *destination* file that we would move the file to, once the user clicks the button.

So, baseForRename might be something like "Lost[1x04] Walkabout".

We were already calculating the originalBasename; it was used as the initialization of baseForRename.  But then it would likely be overwritten.

Now keep it as an instance variable, and don't bother initializing baseForRename; it can be null until we have an actual destination file.

We do some error checking inside getRenamedBasename.  In the case of a severe error, to fail gracefully, we were returning the original filename.  Since this method is supposed to return a basename, it makes a lot more sense to return originalBasename.  Again, though, this is just theoretical; this code should never be hit.

Then, in getDestinationBasename, we had a bug where we were always returning baseForRename.  When rename is not enabled, we should return originalBasename.