Better icon choice for "options"
When I added the implementation of Issue #261, to give users the option of choosing a particular episode, the "status" of a line with options was not distinct from that of a line with one definitive episode found.  Then, I added an "options" status; but the icon I used for it was not intuitive at all.  I just wanted to plug something in and get going.  Now I looked and found a better icon to use.  Replace the previous choice with "circle-green-add".