Replace preferences dialog "save" button with "apply"
This is an intermediate commit.  I want to *add* an apply button.  But to do it in digestible steps, this commit simply *changes* the "save" button to "apply".  In this state, to accept your changes, you have to first apply, and then close.  Obviously that's not good, but will be changed soon.

Also, there's no way for the user to force the prefs to be saved to a file.  "Apply" and "Close" simply do those things.  As of this commit, there really is no "save".