Either "move" or "rename" must be selected
Constrain the preferences dialog so that users cannot de-select both move and rename.  It doesn't make sense to have them both de-selected.

This commit only concerns the preferences dialog.  In the UserPreferences, where the values are stored, we still allow it to be the case that we neither move nor rename the files.  That should also be resolved, before the issue is closed.

Since previous versions allowed users to specify this combination, it's possible that someone has it saved out into their preferences.  Now the preferences dialog will catch that situation, and correct it (by enabling rename), and then updating the preferences, and saving the preferences to disk.