Rework season prefix
We decided that in the preferences dialog, the season prefix will be displayed with double quotes, which obviously do not become part of the file path.  So to go back and forth between the preferences dialog and UserPreferences, we needed to add and subtract the quotes.

But the usage of the quotes has nothing to do with the attribute itself.  It is just about display, for the dialog.

Move all the quoting and unquoting stuff into StringUtils.  Now UserPreferences expects to get the text already stripped and sanitized, and uses it as-is.

Because the field initializes the text to be surrounded by double quotes, we need to strip the double quotes away.  But we were just blindly removing any instance of the double quote character.

Now we will only remove quotes at the beginning and end of the text.  If there are any others, we leave them alone in the first step.  This is implemented in the new StringUtils static method, unquoteString.