Rework replaceIllegalCharacters, use a hash map
Build a mapping between illegal characters and their replacements, and use it in sanitiseTitle.

The point is to expose the mapping, to other parts of the program, and potentially to the user.  We could, in the future, allow users to specify their own preferred substitutions.

Although it's not the motivation for the change, this does make sanitiseTitle much, much faster.  Doing three million iterations of 10-character, randomly generated strings used to take 9.5 - 10.5 seconds, and now it takes 0.7 - 1.5.

But, again, improving something that only took 3 milliseconds to begin with, and making it faster, is not going to be noticeable.  The point of the commit is exposing the mapping.