Add test to measure the speed of sanitiseTitle
The test generates random strings beforehand, into an array.  Then, for the test, it iterates over the array and calls sanitiseTitle on each.  It puts the result into an array, too, which is ignored.  I put them into the array to try to avoid any compiler optimization that might notice that the return value is not being used.

For three million random strings, it usually takes between 10.0 and 10.5 seconds on my laptop.