Disallow "move" and "rename" both being unselected, when reading XML
When the application starts up, it tries to read the user preferences from an XML file.  If no file is found, it instantiates UserPreferences, in a way known to be sound.  But if the file is found, the values in the file are taken and accepted with no verification.

A recent commit added functionality so that the "ignore keywords" string was built upon successful reading of the file, and then moved that call into a new "sanitise" method.  We can add a check there to detect the erroneous situation where neither move nor rename are selected, and correct it, in sanitise().