Create UserPreferences sanitise method
When the application starts up, it tries to read the user preferences from an XML file.  If no file is found, it instantiates UserPreferences, with some hard-coded values.  But if the file is found, the values in the file are taken and accepted with no verification.

A recent commit added functionality so that the "ignore keywords" string was built upon successful reading of the file.  That was inlined into the load() method in the case of successful XML loading, and also added to the constructor.

Upon reflection that there should be more done get the object ready, create a new method, sanitise().  Move the call to buildIgnoredKeywordsString into sanitise(), and call sanitise() from load(), regardless of whether we loaded XML or not.  Then we can remove buildIgnoredKeywordsString from the constructor.