Add checkForWritableDirectory
Unlike the "ensure" methods, does not try to actually create the directory.  Does not even try to create the parent.  Goes as far up the path as necessary to find something that exists, and if that "thing" is a writable directory, we assume we could create the other components, if/when necessary.

Add some testing for these new functionalities.