Don't crash on xstream error
We never wanted to let an xstream error cause the whole program to crash.  If we can't read the prefs.xml file, we abort THAT operation, but we're supposed to just continue on with a default user preferences.

Two things were stopping this:
(1) although we were catching four specific exceptions, we were not catching the one that actually happened (ExceptionInInitializerError).  Rather than add that one, just catch all Exceptions and do the same thing.
(2) when we do fail, don't try to get the singleton instance.  We are trying to CREATE the singleton instance.  This could result in an infinite loop.  Just return null on failure, and UserPreferences will see the null and provide a default, instead.

Conflicts:
	src/main/org/tvrenamer/controller/UserPreferencesPersistence.java