Add EpisodeView
This is most of the implementation I'll need to use it, and I have tested it with many changes to ResultsTable.java.  But I'm doing the commits separately.  This is just EpisodeView and the minimal changes to ResultsTable to allow it to compile.  As of this commit, EpisodeView is still not actually used.