Add GroupedMoveList class
It is a thin wrapper around a HashMap from a String to a "Moves" object.  Just like the previous commit, which introduced the "Moves" class, this is primarily around readability and the lack of it when using Java's util classes inline.

This also allows us to create the addMove method.  This associates a value with a key, by either adding the value to the key's existing Moves, or creating a new Moves, adding the value, and mapping it to the key.  This was tedious to do when using the Map directly.

In order to detect any potential conflicts, when we have a bunch of moves to so, we group them together based on the intended destination.  (Obviously, in the case where there are no conflicts, which we hope will be the normal case, this will mean each group contains exactly one move; but when there are conflicts, some group(s) will contain more.)