Add Moves class
The Moves class is essentially just a collection of FileMover objects; it is a plug-in replacement for List<FileMover>.  This does not really provide functional benefits; I just find Java's collections often clunky, and think having an actual class makes the code much cleaner.

(In addition to simply replacing ugly usages of "List<FileMover>" with "Moves", this also allows a couple of classes to stop importing java.util classes, which I consider a win...)

In the places where I changed a method's signature to use Moves instead of List<FileMover>, also changed all arguments to be final while I was there.